requirements.txt
boto3==1.42.95
pandas==3.0.2
python-dotenv==1.2.2
SQLAlchemy==2.0.49
psycopg2-binary